[CRT] USE CMake for CRT standalone libraries (#14025)
This PR converts Makefile for standalone CRT libraries to a CMake file which is usable in template projects. After this PR we no longer need to have custom CMake scripts to build CRT standalone libraries but instead just include CMakeLists.txt file that is copied in the CRT lib and link the libraries.
Also changed apps/bundle_deploy/Makefile which was a result of this change.