[Refactor] Move `VarUseDefAnalysis` to header file (#14185)
# Motivation
`UndefinedVars` is a frequently used function in our codebase and its implementation relies on `VarUseDefAnalysis` class which is more general, currently we expose `UndefinedVars` in `analysis.h`, but both the definitions of `UndefinedVars` and `VarUseDefAnalysis` resides in `split_host_device.cc`.

This PR moves `VarUseDefAnalysis` class to `analysis.h` so that developers can use it in other files that requires use/def analysis than `split_host_devices.cc`. We create a `var_use_def_analysis.cc` under `src/src/analysis` for the implementations of both `UndefinedVars` and `VarUseDefAnalysis`.

# Notes
We rename `VarUseDefAnalysis` to `VarUseDefAnalyzer`.